chore(structure): move StateBadge to /custom-ui folder and update middleware config to remove references to deprecated assets directory
