+++ packages/tv-ad-admin-next/components/list/order-row.tsx
@@ -1,5 +1,5 @@
+import { StateBadge } from '@/components/custom-ui/state-badge'
 import { Button } from '@/components/ui/button'
-import { StateBadge } from '@/components/ui/state-badge'
 import ArrowRightDownIcon from '@/public/icons/arrow-right-sown.svg'
 import DetailIcon from '@/public/icons/detail.svg'
 import { type OrderRecordForList } from '@/types/order'
